requirements.txt
ydata-profiling==4.12.1
attr==0.3.2
colorama==0.4.6
ConfigParser==7.1.0
cryptography==44.0.0
docutils==0.21.2
filelock==3.16.1
HTMLParser==0.0.2
ipython==8.12.3
ipywidgets==8.1.5
jnius==1.1.0
kagglehub==0.3.6
keyring==25.6.0
pandas==2.2.3
Pillow==11.1.0
protobuf==5.29.3
pyOpenSSL==25.0.0
redis==5.2.1
selenium==4.27.1
Sphinx==8.1.3
thread==2.0.5
undetected_chromedriver==3.5.5
urllib3_secure_extra==0.1.0
webdriver_manager==4.0.2
xmlrpclib==1.0.1
google-api-core==2.24.0
google-auth==2.37.0
google-cloud==0.34.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
duckdb==1.1.3
pytest-8.3.4